backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator",
    "fastapi",
    "passlib[bcrypt]",
    "psycopg2-binary>=2.9.11",
    "pydantic",
    "pyjwt",
    "python-multipart",
    "sqlalchemy>=2.0.45",
    "uvicorn",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
